requirements.txt
Flask>=2.0
python-dotenv>=0.19
Flask-JWT-Extended>=4.4
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
psycopg2-binary>=2.9
Flask-Cors>=3.1
marshmallow>=3.18
Flask-Marshmallow>=0.15
pytest>=7.0
pytest-flask>=1.2
gunicorn
flasgger
